Add native multi-GPU search and reproducible scaling benchmarks
Select all visible CUDA GPUs by default or an explicit --devices list.
Run disjoint denominator batches with device-bound workers, independent
GMP verification, ordered main-thread output, and joined error handling.
Add device listing, per-device metrics, and separate wall-clock timing.
Fix inclusive INT_MAX denominator iteration in the sieve plan.

Add an output-checked single/multi-GPU benchmark and hardware test target.
Validation: 89 host checks with a mocked CUDA backend, also passing under
ASan/UBSan; real host sieve-plan boundary tests; Python syntax checks.
CUDA compilation, kernel execution, and GPU speedup were not tested in
the implementation environment, which has no CUDA toolkit or GPU.

Committed directly to master as requested; no kernel tuning changes.

+++ src/sieve_plan.cpp
@@ -68,7 +68,8 @@ SievePlan::SievePlan(long long numerator_bound, DenominatorRange range)
     }
 
     selected_primes.resize(static_cast<size_t>(range.count()) * kPrimeCount);
-    for (int denominator = range.first;
+    // The inclusive endpoint can be INT_MAX; increment a wider cursor.
+    for (long long denominator = range.first;
          denominator <= range.last; ++denominator) {
         int selected_count = 0;
         size_t row = static_cast<size_t>(
